[CHANGE] Template tags combined into one file
