+++ aa_forum/templates/aa_forum/bundles/aa-forum-admin-js.html
@@ -1,3 +1,3 @@
-{% load aa_forum_versioned_static %}
+{% load aa_forum %}
 
 <script src="{% aa_forum_static 'aa_forum/javascript/aa-forum-admin.min.js' %}"></script>
