+++ aa_forum/templates/aa_forum/bundles/aa-forum-ckeditor-js.html
@@ -1,3 +1,3 @@
-{% load aa_forum_versioned_static %}
+{% load aa_forum %}
 
 <script src="{% aa_forum_static 'aa_forum/javascript/aa-forum-ckeditor.min.js' %}"></script>
