+++ aa_forum/templates/aa_forum/bundles/aa-forum-css.html
@@ -1,3 +1,3 @@
-{% load aa_forum_versioned_static %}
+{% load aa_forum %}
 
 <link rel="stylesheet" href="{% aa_forum_static 'aa_forum/css/aa-forum.min.css' %}">
