+++ aa_forum/templates/aa_forum/partials/breadcrumb.html
@@ -1,4 +1,4 @@
-{% load aa_forum_template_variables %}
+{% load aa_forum %}
 {% load i18n %}
 
 <div class="aa-forum-breadcrumb px-3 py-2">
@@ -13,8 +13,8 @@
             <li>» <a href="{% url 'aa_forum:forum_index' %}#category-{{ category.id }}">{{ category.name }}</a></li>
 
             {% if board %}
-                {% set_template_variable category_slug = category.slug %}
-                {% set_template_variable board_slug = board.slug %}
+                {% aa_forum_template_variable category_slug = category.slug %}
+                {% aa_forum_template_variable board_slug = board.slug %}
 
                 {% if board.parent_board %}
                     <li>» <a href="{{ board.parent_board.get_absolute_url }}">{{ board.parent_board.name }}</a></li>
