+++ aa_forum/templates/aa_forum/partials/forum/board/pagination.html
@@ -1,9 +1,9 @@
-{% load aa_forum_template_variables %}
+{% load aa_forum %}
 {% load i18n %}
 
 {% if page_obj.paginator.num_pages > 1 %}
     <div class="col-4 offset-4">
-        {% set_template_variable category = board.category %}
+        {% aa_forum_template_variable category = board.category %}
 
         <nav aria-label="{% translate 'Message pagination' %}">
             <ul class="pagination pagination-sm">
