+++ aa_forum/templates/aa_forum/partials/forum/board/topic.html
@@ -1,7 +1,6 @@
 {% load i18n %}
 {% load static %}
-{% load aa_forum_datetime %}
-{% load aa_forum_template_variables %}
+{% load aa_forum %}
 
 <div class="aa-forum-topic-row {% if topic.is_locked %}aa-forum-topic-locked{% endif %} {% if topic.is_sticky %}aa-forum-topic-sticky{% endif %} row">
     <div class="aa-forum-topic">
@@ -45,7 +44,7 @@
                         </span>
                     {% endif %}
 
-                    {% translate "Started by" %} {{ topic.first_message.user_created|main_character_name }}
+                    {% translate "Started by" %} {{ topic.first_message.user_created|aa_forum_main_character_name }}
                 </span>
             </p>
         </div>
@@ -66,14 +65,14 @@
 
         <div class="aa-forum-topic-last-post hidden-xs">
             <span class="small">
-                {% translate "Last post" %} {{ topic.last_message.time_posted|forum_time }}
+                {% translate "Last post" %} {{ topic.last_message.time_posted|aa_forum_time }}
 
                 <a href="{% url 'aa_forum:forum_message' topic.board.category.slug topic.board.slug topic.slug topic.last_message_id %}" title="{% translate 'Go to last message' %}" style="margin-left: 0.5rem;">
                     <i class="fas fa-sign-in-alt"></i>
                 </a>
 
                 <br>
-                by {{ topic.last_message.user_created|main_character_name }}
+                by {{ topic.last_message.user_created|aa_forum_main_character_name }}
             </span>
         </div>
