+++ aa_forum/templates/aa_forum/partials/forum/topic/message.html
@@ -1,11 +1,8 @@
 {% load i18n %}
-{% load aa_forum_datetime %}
-{% load aa_forum_search %}
-{% load aa_forum_template_variables %}
-{% load aa_forum_user %}
+{% load aa_forum %}
 
 <div id="message-{{ message.id }}" class="card card-default mb-3">
-    {% set_template_variable message_author = message.user_created %}
+    {% aa_forum_template_variable message_author = message.user_created %}
 
     <div class="card-body">
         <div class="aa-forum-message">
@@ -27,7 +24,7 @@
 
                             {{ topic.subject }}
                         </a><br>
-                        <small>« <b>{% translate "on:" %}</b> {{ message.time_posted|forum_time }} »</small>
+                        <small>« <b>{% translate "on:" %}</b> {{ message.time_posted|aa_forum_time }} »</small>
                     </div>
 
                     <div class="col-md-6">
@@ -70,7 +67,7 @@
                 <div class="aa-forum-message-body">
                     <div class="aa-forum-message-body-inner">
                         {% if search_term %}
-                            {{ message.message|highlight_search_term:search_term }}
+                            {{ message.message|aa_forum_highlight_search_term:search_term }}
                         {% else %}
                             {{ message.message|safe }}
                         {% endif %}
@@ -78,7 +75,7 @@
 
                     {% if message.user_updated %}
                         <div class="aa-forum-message-body-last-edited text-muted small">
-                            « {% translate "Last modified:" %} {{ message.time_modified|forum_time }} by {{ message.user_updated|main_character_name }} »
+                            « {% translate "Last modified:" %} {{ message.time_modified|aa_forum_time }} by {{ message.user_updated|aa_forum_main_character_name }} »
                         </div>
                     {% endif %}
