+++ aa_forum/templates/aa_forum/partials/forum/topic/pagination.html
@@ -1,11 +1,11 @@
-{% load aa_forum_template_variables %}
+{% load aa_forum %}
 {% load i18n %}
 
 {% if page_obj.paginator.num_pages > 1 %}
     <div class="col-4 offset-4">
-        {% set_template_variable topic_first_message = topic.first_message %}
-        {% set_template_variable topic_first_message_board = topic_first_message.topic.board %}
-        {% set_template_variable topic_first_message_category = topic_first_message_board.category %}
+        {% aa_forum_template_variable topic_first_message = topic.first_message %}
+        {% aa_forum_template_variable topic_first_message_board = topic_first_message.topic.board %}
+        {% aa_forum_template_variable topic_first_message_category = topic_first_message_board.category %}
 
         <nav aria-label="{% translate 'Message pagination' %}">
             <ul class="pagination pagination-sm">
