+++ aa_forum/templates/aa_forum/partials/menu/menu-user.html
@@ -1,6 +1,6 @@
 {% load i18n %}
 {% load navactive %}
-{% load aa_forum_template_variables %}
+{% load aa_forum %}
 
 <li class="nav-item">
     <a class="nav-link {% navactive request 'aa_forum:profile_index' %}" href="{% url 'aa_forum:profile_index' %}">
