+++ aa_forum/templates/aa_forum/partials/personal-messages/inbox/messages.html
@@ -1,7 +1,6 @@
 {% load i18n %}
 {% load evelinks %}
-{% load aa_forum_datetime %}
-{% load aa_forum_user %}
+{% load aa_forum %}
 
 <div class="col-sm-8 col-md-9 col-lg-10 order-sm-first mb-3">
     <div class="card card-default mb-3 card-aa-forum-personal-messages-list">
@@ -31,21 +30,21 @@
                     <div id="aa-forum-personal-message-id-{{ message.pk }}">
                         <div class="card-aa-forum-personal-messages-item {% if not message.is_read %}card-aa-forum-personal-messages-item-unread{% endif %}">
                             <div class="aa-forum-personal-message-date">
-                                {{ message.time_sent|forum_time }}
+                                {{ message.time_sent|aa_forum_time }}
                             </div>
 
                             <div class="aa-forum-personal-message-sender">
-                                {% with sender_maincharacter_id=message.sender|main_character_id %}
+                                {% with sender_maincharacter_id=message.sender|aa_forum_main_character_id %}
                                     <img
                                         class="aa-forum-character-portrait img img-rounded img-last-post-avatar"
                                         src="{{ sender_maincharacter_id|character_portrait_url:32 }}"
-                                        alt="{{ message.sender|main_character_name }}"
+                                        alt="{{ message.sender|aa_forum_main_character_name }}"
                                         width="32"
                                         height="32"
                                     >
                                 {% endwith %}
 
-                                {{ message.sender|main_character_name  }}
+                                {{ message.sender|aa_forum_main_character_name }}
                             </div>
 
                             <div class="aa-forum-personal-message-subject">
