+++ aa_forum/templates/aa_forum/partials/personal-messages/sidebar.html
@@ -1,6 +1,6 @@
 {% load i18n %}
 {% load navactive %}
-{% load aa_forum_template_variables %}
+{% load aa_forum %}
 
 <div class="col-sm-4 col-md-3 col-lg-2 order-sm-last mb-3">
     <div class="card card-default">
