+++ aa_forum/templates/aa_forum/partials/search/search-result.html
@@ -1,4 +1,4 @@
-{% load aa_forum_search %}
+{% load aa_forum %}
 
 <div class="aa-forum-search-result">
     <div class="aa-forum-search-result-inner">
